+++ internal/core/Composite.go
@@ -2,8 +2,8 @@ package core
 
 type IComposite interface {
 	Node
-	GetChildCount() int
-	GetChild(index int) Node
+	GetChildCount() int64
+	GetChild(index int64) Node
 	AddChild(child Node)
 }
 
@@ -17,11 +17,11 @@ func (c *Composite) GetCategory() Category {
 	return COMPOSITE
 }
 
-func (c *Composite) GetChildCount() int {
-	return len(c.children)
+func (c *Composite) GetChildCount() int64 {
+	return int64(len(c.children))
 }
 
-func (c *Composite) GetChild(index int) Node {
+func (c *Composite) GetChild(index int64) Node {
 	return c.children[index]
 }
